+++ main.js
@@ -1,5 +1,8 @@
-import { bandwidthCalc,streaming } from "./src/bandwidth-calc.js"
-import { calculator } from "../bytes-and-bits-converter/script.js"
+// import { bandwidthCalc,streaming } from "./src/bandwidth-calc.js"
+// import { calculator } from "../bytes-and-bits-converter/script.js"
+const { bandwidthCalc,streaming } = require('./src/bandwidth-calc')
+const calculator = require('bytes-and-bits-converter')
 
-export { bandwidthCalc }
-export { calculator,streaming }
+// export { bandwidthCalc }
+// export { calculator,streaming }
+module.exports = { bandwidthCalc,calculator,streaming }